Copy terraformer binary after tf and plugins
When developing on terraformer, pushing dev builds can take quite long,
as every change in the terraformer binary will result in pushes for the
terraform and terraform plugins layer. By copying the terraformer binary
last in the image build, the unchanged terraform and plugin layers don't
have to be pushed again.